+++ rover-server/internal/controller/apispecification_lint_test.go
@@ -5,13 +5,10 @@
 package controller
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
 	"io"
 	"net/http"
-	"net/http/httptest"
 
 	"github.com/stretchr/testify/mock"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -22,7 +19,6 @@ import (
 	"github.com/telekom/controlplane/common-server/pkg/store"
 	fileApi "github.com/telekom/controlplane/file-manager/api"
 	"github.com/telekom/controlplane/rover-server/internal/api"
-	"github.com/telekom/controlplane/rover-server/internal/config"
 	"github.com/telekom/controlplane/rover-server/internal/oaslint"
 	s "github.com/telekom/controlplane/rover-server/pkg/store"
 	"github.com/telekom/controlplane/rover-server/test/mocks"
@@ -32,295 +28,14 @@ import (
 	. "github.com/onsi/gomega"
 )
 
-var _ = Describe("Linting helpers", func() {
-	Describe("prepareLinting", func() {
-		var linter *apiLinterImpl
-
-		BeforeEach(func() {
-			linter = &apiLinterImpl{}
-		})
-
-		It("should skip linting for category-whitelisted basepath", func() {
-			lintCfg := &apiv1.LintingConfig{
-				WhitelistedBasepaths: []string{"/eni/internal/v1"},
-			}
-			apiSpec := &roverv1.ApiSpecification{
-				Spec: roverv1.ApiSpecificationSpec{BasePath: "/eni/internal/v1"},
-			}
-
-			result := linter.prepareLinting(lintCfg, apiSpec)
-			Expect(result).To(BeFalse())
-			Expect(apiSpec.Spec.Lint).ToNot(BeNil())
-			Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
-			Expect(apiSpec.Spec.Lint.Message).To(ContainSubstring("whitelisted"))
-		})
-
-		It("should require linting when basepath is not whitelisted", func() {
-			lintCfg := &apiv1.LintingConfig{}
-			apiSpec := &roverv1.ApiSpecification{
-				Spec: roverv1.ApiSpecificationSpec{BasePath: "/eni/test/v1"},
-			}
-
-			result := linter.prepareLinting(lintCfg, apiSpec)
-			Expect(result).To(BeTrue())
-			Expect(apiSpec.Spec.Lint).To(BeNil())
-		})
-	})
-
-	Describe("Lint", func() {
-		var (
-			lintCtx      context.Context
-			linterServer *httptest.Server
-			linter       ApiLinter
-			apiSpec      *roverv1.ApiSpecification
-			category     *apiv1.ApiCategory
-			specBytes    io.Reader
-		)
-
-		newCategory := func(mode apiv1.LintingMode) *apiv1.ApiCategory {
-			return &apiv1.ApiCategory{
-				ObjectMeta: metav1.ObjectMeta{Name: "test-cat"},
-				Spec: apiv1.ApiCategorySpec{
-					LabelValue: "test-cat",
-					Active:     true,
-					Linting: &apiv1.LintingConfig{
-						Mode:    mode,
-						Ruleset: "default",
-					},
-				},
-			}
-		}
-
-		startLinterServer := func(errors int) *httptest.Server {
-			return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
-				resp := map[string]any{
-					"id":            "scan-test",
-					"createdAt":     "2025-01-01T00:00:00Z",
-					"ruleset":       map[string]any{"name": "default", "hash": "abc"},
-					"info":          map[string]any{"errors": errors, "warnings": 0, "infos": 0, "hints": 0},
-					"linterVersion": "1.0.0",
-				}
-				w.Header().Set("Content-Type", "application/json")
-				json.NewEncoder(w).Encode(resp) //nolint:errcheck // test helper
-			}))
-		}
-
-		BeforeEach(func() {
-			lintCtx = context.Background()
-			specBytes = bytes.NewBuffer([]byte("openapi: '3.0.0'"))
-			apiSpec = &roverv1.ApiSpecification{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "test-spec",
-					Namespace: "env--ns",
-				},
-				Spec: roverv1.ApiSpecificationSpec{
-					BasePath: "/test/api/v1",
-					Category: "test-cat",
-					Hash:     "new-hash",
-				},
-			}
-		})
-
-		AfterEach(func() {
-			if linterServer != nil {
-				linterServer.Close()
-			}
-		})
-
-		It("should skip when category is nil", func() {
-			linter = &apiLinterImpl{url: "http://linter"}
-			outcome, err := linter.Lint(lintCtx, apiSpec, nil, specBytes)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(outcome).To(Equal(LintSkipped))
-		})
-
-		It("should skip when mode is None", func() {
-			linter = &apiLinterImpl{url: "http://linter"}
-			category = newCategory(apiv1.LintingModeNone)
-			outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(outcome).To(Equal(LintSkipped))
-		})
-
-		It("should skip when linter URL is empty", func() {
-			linter = NewApiLinter(config.OasLintingConfig{URL: ""})
-			category = newCategory(apiv1.LintingModeWarn)
-			outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(outcome).To(Equal(LintSkipped))
-		})
-
-		Context("when linting passes", func() {
-			BeforeEach(func() {
-				linterServer = startLinterServer(0)
-			})
-
-			It("should return LintCompleted in Block mode", func() {
-				linter = &apiLinterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
-				category = newCategory(apiv1.LintingModeBlock)
-				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outcome).To(Equal(LintCompleted))
-				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
-				Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
-			})
-
-			It("should return LintCompleted in Warn mode", func() {
-				linter = &apiLinterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
-				category = newCategory(apiv1.LintingModeWarn)
-				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outcome).To(Equal(LintCompleted))
-				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
-				Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
-			})
-		})
-
-		Context("when linting fails (spec has errors)", func() {
-			BeforeEach(func() {
-				linterServer = startLinterServer(3)
-			})
-
-			It("should return LintBlocked without error in Block mode", func() {
-				linter = &apiLinterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
-				category = newCategory(apiv1.LintingModeBlock)
-				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outcome).To(Equal(LintBlocked))
-				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
-				Expect(apiSpec.Spec.Lint.Passed).To(BeFalse())
-			})
-
-			It("should return LintCompleted without error in Warn mode", func() {
-				linter = &apiLinterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
-				category = newCategory(apiv1.LintingModeWarn)
-				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-				Expect(err).ToNot(HaveOccurred())
-				Expect(outcome).To(Equal(LintCompleted))
-				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
-				Expect(apiSpec.Spec.Lint.Passed).To(BeFalse())
-			})
-		})
-
-		Context("when linter API is unreachable", func() {
-			It("should return error without persisting", func() {
-				linter = &apiLinterImpl{url: "http://localhost:1", httpClient: &http.Client{}}
-				category = newCategory(apiv1.LintingModeBlock)
-				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
-				Expect(err).To(HaveOccurred())
-				Expect(outcome).To(Equal(LintCompleted))
-			})
-		})
-	})
-
-	Describe("buildLintResult", func() {
-		It("should substitute {{.LinterId}} in dashboardURL", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
-				errorMessageTemplate: "failed",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   true,
-				Reason:   "ok",
-				Ruleset:  "default",
-				LinterId: "scan-abc-123",
-			})
-			Expect(result.DashboardURL).To(Equal("https://linter.example.com/scans/scan-abc-123"))
-		})
-
-		It("should substitute {{.RulesetName}} in dashboardURL", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "https://editor.example.com/tooling/oas-editor?ruleset={{.RulesetName}}",
-				errorMessageTemplate: "failed",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   false,
-				Reason:   "errors found",
-				Ruleset:  "gapi",
-				LinterId: "scan-1",
-			})
-			Expect(result.DashboardURL).To(Equal("https://editor.example.com/tooling/oas-editor?ruleset=gapi"))
-		})
-
-		It("should substitute both placeholders in dashboardURL", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}?ruleset={{.RulesetName}}",
-				errorMessageTemplate: "failed",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   false,
-				Reason:   "errors",
-				Ruleset:  "strict",
-				LinterId: "id-42",
-			})
-			Expect(result.DashboardURL).To(Equal("https://linter.example.com/scans/id-42?ruleset=strict"))
-		})
-
-		It("should leave dashboardURL empty when not configured", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "",
-				errorMessageTemplate: "failed",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   true,
-				Reason:   "ok",
-				Ruleset:  "default",
-				LinterId: "scan-1",
-			})
-			Expect(result.DashboardURL).To(BeEmpty())
-		})
-
-		It("should substitute {{.RulesetName}} in errorMessage when linting fails", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "",
-				errorMessageTemplate: "Linting failed for {{.RulesetName}} ruleset.",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:  false,
-				Reason:  "errors found",
-				Ruleset: "gapi",
-			})
-			Expect(result.Message).To(Equal("Linting failed for gapi ruleset."))
-		})
-
-		It("should substitute {{.DashboardURL}} in errorMessage when linting fails", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
-				errorMessageTemplate: "Linting failed for {{.RulesetName}} ruleset. {{.DashboardURL}}",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   false,
-				Reason:   "errors found",
-				Ruleset:  "gapi",
-				LinterId: "scan-99",
-			})
-			Expect(result.Message).To(Equal("Linting failed for gapi ruleset. https://linter.example.com/scans/scan-99"))
-		})
-
-		It("should not override message when linting passes", func() {
-			l := &apiLinterImpl{
-				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
-				errorMessageTemplate: "This should not appear",
-			}
-			result := l.buildLintResult(&oaslint.LintResult{
-				Passed:   true,
-				Reason:   "all good",
-				Ruleset:  "default",
-				LinterId: "scan-1",
-			})
-			Expect(result.Message).To(Equal("all good"))
-		})
-	})
-})
-
-// mockLinter is a simple test double for ApiLinter that records whether Lint was called.
+// mockLinter is a simple test double for oaslint.Linter that records whether Lint was called.
 type mockLinter struct {
 	called  bool
-	outcome LintOutcome
+	outcome oaslint.Outcome
 	err     error
 }
 
-func (m *mockLinter) Lint(_ context.Context, apiSpec *roverv1.ApiSpecification, _ *apiv1.ApiCategory, _ io.Reader) (LintOutcome, error) {
+func (m *mockLinter) Lint(_ context.Context, apiSpec *roverv1.ApiSpecification, _ *apiv1.ApiCategory, _ io.Reader) (oaslint.Outcome, error) {
 	m.called = true
 	if apiSpec.Spec.Lint == nil {
 		apiSpec.Spec.Lint = &roverv1.LintResult{Passed: true, Message: "mock lint ran"}
@@ -336,6 +51,7 @@ var _ = Describe("Linter unreachable error propagation to user", func() {
 
 	newCategoryWithMode := func(mode apiv1.LintingMode) *apiv1.ApiCategory {
 		return &apiv1.ApiCategory{
+			ObjectMeta: metav1.ObjectMeta{Name: "test-cat"},
 			Spec: apiv1.ApiCategorySpec{
 				LabelValue: "test-cat",
 				Active:     true,
@@ -381,14 +97,13 @@ var _ = Describe("Linter unreachable error propagation to user", func() {
 				}}, nil)
 
 			specStore := mocks.NewMockObjectStore[*roverv1.ApiSpecification](GinkgoT())
-			// lintSpec calls Get for hash dedup; return not-found so linting proceeds.
 			specStore.EXPECT().Get(mock.Anything, mock.Anything, mock.Anything).Return(
 				nil, problems.NotFound())
 
 			ctrl = &ApiSpecificationController{
 				stores: &s.Stores{APICategoryStore: categoryStore},
 				Store:  specStore,
-				Linter: &mockLinter{outcome: LintCompleted, err: fmt.Errorf("connection refused")},
+				Linter: &mockLinter{outcome: oaslint.Completed, err: fmt.Errorf("connection refused")},
 			}
 
 			_, err := ctrl.Update(testCtx, "eni--hyperion--test-api-v1", newUpdateRequest())
@@ -410,19 +125,17 @@ var _ = Describe("Linter unreachable error propagation to user", func() {
 				}}, nil)
 
 			specStore := mocks.NewMockObjectStore[*roverv1.ApiSpecification](GinkgoT())
-			// Update calls Get for hash dedup; return not-found so it proceeds.
 			specStore.EXPECT().Get(mock.Anything, mock.Anything, mock.Anything).Return(
 				nil, problems.NotFound())
 			specStore.EXPECT().CreateOrReplace(mock.Anything, mock.Anything).Return(nil)
 
-			// Mock the file manager for the upload that happens after linting passes
 			mockFileManager.EXPECT().UploadFile(mock.Anything, mock.Anything, mock.Anything, mock.Anything).Return(
 				&fileApi.FileUploadResponse{FileHash: "abc", FileId: "test-id", ContentType: "application/yaml"}, nil)
 
 			ctrl = &ApiSpecificationController{
 				stores: &s.Stores{APICategoryStore: categoryStore},
 				Store:  specStore,
-				Linter: &mockLinter{outcome: LintCompleted, err: fmt.Errorf("connection refused")},
+				Linter: &mockLinter{outcome: oaslint.Completed, err: fmt.Errorf("connection refused")},
 			}
 
 			_, err := ctrl.Update(testCtx, "eni--hyperion--test-api-v1", newUpdateRequest())
@@ -439,7 +152,7 @@ var _ = Describe("Linter unreachable error propagation to user", func() {
 		})
 	})
 
-	Context("when linter returns LintBlocked", func() {
+	Context("when linter returns Blocked", func() {
 		It("should return BadRequest to the user", func() {
 			categoryStore := mocks.NewMockObjectStore[*apiv1.ApiCategory](GinkgoT())
 			categoryStore.EXPECT().List(mock.Anything, mock.Anything).Return(
@@ -448,14 +161,13 @@ var _ = Describe("Linter unreachable error propagation to user", func() {
 				}}, nil)
 
 			specStore := mocks.NewMockObjectStore[*roverv1.ApiSpecification](GinkgoT())
-			// Update calls Get for hash dedup; return not-found so it proceeds.
 			specStore.EXPECT().Get(mock.Anything, mock.Anything, mock.Anything).Return(
 				nil, problems.NotFound())
 
 			ctrl = &ApiSpecificationController{
 				stores: &s.Stores{APICategoryStore: categoryStore},
 				Store:  specStore,
-				Linter: &mockLinter{outcome: LintBlocked, err: nil},
+				Linter: &mockLinter{outcome: oaslint.Blocked, err: nil},
 			}
 
 			_, err := ctrl.Update(testCtx, "eni--hyperion--test-api-v1", newUpdateRequest())
