+++ rover-server/internal/oaslint/linter_test.go
@@ -1,0 +1,304 @@
+// Copyright 2025 Deutsche Telekom IT GmbH
+//
+// SPDX-License-Identifier: Apache-2.0
+
+package oaslint
+
+import (
+	"bytes"
+	"context"
+	"encoding/json"
+	"io"
+	"net/http"
+	"net/http/httptest"
+
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+
+	apiv1 "github.com/telekom/controlplane/api/api/v1"
+	"github.com/telekom/controlplane/rover-server/internal/config"
+	roverv1 "github.com/telekom/controlplane/rover/api/v1"
+
+	. "github.com/onsi/ginkgo/v2"
+	. "github.com/onsi/gomega"
+)
+
+var _ = Describe("Linting helpers", func() {
+	Describe("prepareLinting", func() {
+		var linter *linterImpl
+
+		BeforeEach(func() {
+			linter = &linterImpl{}
+		})
+
+		It("should skip linting for category-whitelisted basepath", func() {
+			lintCfg := &apiv1.LintingConfig{
+				WhitelistedBasepaths: []string{"/eni/internal/v1"},
+			}
+			apiSpec := &roverv1.ApiSpecification{
+				Spec: roverv1.ApiSpecificationSpec{BasePath: "/eni/internal/v1"},
+			}
+
+			result := linter.prepareLinting(lintCfg, apiSpec)
+			Expect(result).To(BeFalse())
+			Expect(apiSpec.Spec.Lint).ToNot(BeNil())
+			Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
+			Expect(apiSpec.Spec.Lint.Message).To(ContainSubstring("whitelisted"))
+		})
+
+		It("should require linting when basepath is not whitelisted", func() {
+			lintCfg := &apiv1.LintingConfig{}
+			apiSpec := &roverv1.ApiSpecification{
+				Spec: roverv1.ApiSpecificationSpec{BasePath: "/eni/test/v1"},
+			}
+
+			result := linter.prepareLinting(lintCfg, apiSpec)
+			Expect(result).To(BeTrue())
+			Expect(apiSpec.Spec.Lint).To(BeNil())
+		})
+	})
+
+	Describe("Lint", func() {
+		var (
+			lintCtx      context.Context
+			linterServer *httptest.Server
+			linter       Linter
+			apiSpec      *roverv1.ApiSpecification
+			category     *apiv1.ApiCategory
+			specBytes    io.Reader
+		)
+
+		newCategory := func(mode apiv1.LintingMode) *apiv1.ApiCategory {
+			return &apiv1.ApiCategory{
+				ObjectMeta: metav1.ObjectMeta{Name: "test-cat"},
+				Spec: apiv1.ApiCategorySpec{
+					LabelValue: "test-cat",
+					Active:     true,
+					Linting: &apiv1.LintingConfig{
+						Mode:    mode,
+						Ruleset: "default",
+					},
+				},
+			}
+		}
+
+		startLinterServer := func(errors int) *httptest.Server {
+			return httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, _ *http.Request) {
+				resp := map[string]any{
+					"id":            "scan-test",
+					"createdAt":     "2025-01-01T00:00:00Z",
+					"ruleset":       map[string]any{"name": "default", "hash": "abc"},
+					"info":          map[string]any{"errors": errors, "warnings": 0, "infos": 0, "hints": 0},
+					"linterVersion": "1.0.0",
+				}
+				w.Header().Set("Content-Type", "application/json")
+				json.NewEncoder(w).Encode(resp) //nolint:errcheck // test helper
+			}))
+		}
+
+		BeforeEach(func() {
+			lintCtx = context.Background()
+			specBytes = bytes.NewBuffer([]byte("openapi: '3.0.0'"))
+			apiSpec = &roverv1.ApiSpecification{
+				ObjectMeta: metav1.ObjectMeta{
+					Name:      "test-spec",
+					Namespace: "env--ns",
+				},
+				Spec: roverv1.ApiSpecificationSpec{
+					BasePath: "/test/api/v1",
+					Category: "test-cat",
+					Hash:     "new-hash",
+				},
+			}
+		})
+
+		AfterEach(func() {
+			if linterServer != nil {
+				linterServer.Close()
+			}
+		})
+
+		It("should skip when category is nil", func() {
+			linter = &linterImpl{url: "http://linter"}
+			outcome, err := linter.Lint(lintCtx, apiSpec, nil, specBytes)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(outcome).To(Equal(Skipped))
+		})
+
+		It("should skip when mode is None", func() {
+			linter = &linterImpl{url: "http://linter"}
+			category = newCategory(apiv1.LintingModeNone)
+			outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(outcome).To(Equal(Skipped))
+		})
+
+		It("should skip when linter URL is empty", func() {
+			linter = NewLinter(config.OasLintingConfig{URL: ""})
+			category = newCategory(apiv1.LintingModeWarn)
+			outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+			Expect(err).ToNot(HaveOccurred())
+			Expect(outcome).To(Equal(Skipped))
+		})
+
+		Context("when linting passes", func() {
+			BeforeEach(func() {
+				linterServer = startLinterServer(0)
+			})
+
+			It("should return Completed in Block mode", func() {
+				linter = &linterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
+				category = newCategory(apiv1.LintingModeBlock)
+				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+				Expect(err).ToNot(HaveOccurred())
+				Expect(outcome).To(Equal(Completed))
+				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
+				Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
+			})
+
+			It("should return Completed in Warn mode", func() {
+				linter = &linterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
+				category = newCategory(apiv1.LintingModeWarn)
+				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+				Expect(err).ToNot(HaveOccurred())
+				Expect(outcome).To(Equal(Completed))
+				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
+				Expect(apiSpec.Spec.Lint.Passed).To(BeTrue())
+			})
+		})
+
+		Context("when linting fails (spec has errors)", func() {
+			BeforeEach(func() {
+				linterServer = startLinterServer(3)
+			})
+
+			It("should return Blocked without error in Block mode", func() {
+				linter = &linterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
+				category = newCategory(apiv1.LintingModeBlock)
+				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+				Expect(err).ToNot(HaveOccurred())
+				Expect(outcome).To(Equal(Blocked))
+				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
+				Expect(apiSpec.Spec.Lint.Passed).To(BeFalse())
+			})
+
+			It("should return Completed without error in Warn mode", func() {
+				linter = &linterImpl{url: linterServer.URL, httpClient: linterServer.Client()}
+				category = newCategory(apiv1.LintingModeWarn)
+				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+				Expect(err).ToNot(HaveOccurred())
+				Expect(outcome).To(Equal(Completed))
+				Expect(apiSpec.Spec.Lint).ToNot(BeNil())
+				Expect(apiSpec.Spec.Lint.Passed).To(BeFalse())
+			})
+		})
+
+		Context("when linter API is unreachable", func() {
+			It("should return error without persisting", func() {
+				linter = &linterImpl{url: "http://localhost:1", httpClient: &http.Client{}}
+				category = newCategory(apiv1.LintingModeBlock)
+				outcome, err := linter.Lint(lintCtx, apiSpec, category, specBytes)
+				Expect(err).To(HaveOccurred())
+				Expect(outcome).To(Equal(Completed))
+			})
+		})
+	})
+
+	Describe("buildLintResult", func() {
+		It("should substitute {{.LinterId}} in dashboardURL", func() {
+			l := &linterImpl{
+				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
+				errorMessageTemplate: "failed",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   true,
+				Reason:   "ok",
+				Ruleset:  "default",
+				LinterId: "scan-abc-123",
+			})
+			Expect(result.DashboardURL).To(Equal("https://linter.example.com/scans/scan-abc-123"))
+		})
+
+		It("should substitute {{.RulesetName}} in dashboardURL", func() {
+			l := &linterImpl{
+				dashboardURL:         "https://editor.example.com/tooling/oas-editor?ruleset={{.RulesetName}}",
+				errorMessageTemplate: "failed",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   false,
+				Reason:   "errors found",
+				Ruleset:  "gapi",
+				LinterId: "scan-1",
+			})
+			Expect(result.DashboardURL).To(Equal("https://editor.example.com/tooling/oas-editor?ruleset=gapi"))
+		})
+
+		It("should substitute both placeholders in dashboardURL", func() {
+			l := &linterImpl{
+				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}?ruleset={{.RulesetName}}",
+				errorMessageTemplate: "failed",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   false,
+				Reason:   "errors",
+				Ruleset:  "strict",
+				LinterId: "id-42",
+			})
+			Expect(result.DashboardURL).To(Equal("https://linter.example.com/scans/id-42?ruleset=strict"))
+		})
+
+		It("should leave dashboardURL empty when not configured", func() {
+			l := &linterImpl{
+				dashboardURL:         "",
+				errorMessageTemplate: "failed",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   true,
+				Reason:   "ok",
+				Ruleset:  "default",
+				LinterId: "scan-1",
+			})
+			Expect(result.DashboardURL).To(BeEmpty())
+		})
+
+		It("should substitute {{.RulesetName}} in errorMessage when linting fails", func() {
+			l := &linterImpl{
+				dashboardURL:         "",
+				errorMessageTemplate: "Linting failed for {{.RulesetName}} ruleset.",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:  false,
+				Reason:  "errors found",
+				Ruleset: "gapi",
+			})
+			Expect(result.Message).To(Equal("Linting failed for gapi ruleset."))
+		})
+
+		It("should substitute {{.DashboardURL}} in errorMessage when linting fails", func() {
+			l := &linterImpl{
+				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
+				errorMessageTemplate: "Linting failed for {{.RulesetName}} ruleset. {{.DashboardURL}}",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   false,
+				Reason:   "errors found",
+				Ruleset:  "gapi",
+				LinterId: "scan-99",
+			})
+			Expect(result.Message).To(Equal("Linting failed for gapi ruleset. https://linter.example.com/scans/scan-99"))
+		})
+
+		It("should not override message when linting passes", func() {
+			l := &linterImpl{
+				dashboardURL:         "https://linter.example.com/scans/{{.LinterId}}",
+				errorMessageTemplate: "This should not appear",
+			}
+			result := l.buildLintResult(&scanResult{
+				Passed:   true,
+				Reason:   "all good",
+				Ruleset:  "default",
+				LinterId: "scan-1",
+			})
+			Expect(result.Message).To(Equal("all good"))
+		})
+	})
+})
